#12 refactor: Token 처리 함수 리팩토링 및 병합
