+++ app/src/main/java/com/gildongmu/ddu_ru_mobile/MainActivity.kt
@@ -3,7 +3,7 @@ package com.gildongmu.ddu_ru_mobile
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
-import com.gildongmu.ddu_ru_mobile.ui.LoginActivity
+import com.gildongmu.ddu_ru_mobile.ui.auth.LoginActivity
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
