--- app/src/main/java/com/gildongmu/ddu_ru_mobile/data/model/request/LoginRequest.kt
@@ -1,5 +1,0 @@
-package com.gildongmu.ddu_ru_mobile.data.model.request
-
-data class LoginRequest(
-    val idToken : String
-)
