--- app/src/main/java/com/gildongmu/ddu_ru_mobile/model/KakaoUser.kt
@@ -1,10 +1,0 @@
-package com.gildongmu.ddu_ru_mobile.model
-
-data class KakaoLoginRequest(
-    val idToken: String
-)
-
-data class KakaoLoginResponse(
-    val accessToken: String,
-    val refreshToken: String
-)
