+++ app/src/main/java/com/gildongmu/ddu_ru_mobile/model/auth/request/LoginRequest.kt
@@ -1,0 +1,5 @@
+package com.gildongmu.ddu_ru_mobile.model.auth.request
+
+data class LoginRequest(
+    val idToken : String
+)
