+++ app/src/main/java/com/gildongmu/ddu_ru_mobile/model/auth/response/LoginResponse.kt
@@ -1,4 +1,4 @@
-package com.gildongmu.ddu_ru_mobile.data.model.response
+package com.gildongmu.ddu_ru_mobile.model.auth.response
 
 data class LoginResponse(
     val accessToken: String,
