+++ app/src/main/java/com/gildongmu/ddu_ru_mobile/network/NetworkModule.kt
@@ -2,6 +2,7 @@ package com.gildongmu.ddu_ru_mobile.network
 
 import android.content.Context
 import com.gildongmu.ddu_ru_mobile.BuildConfig
+import com.gildongmu.ddu_ru_mobile.network.api.auth.AuthService
 import retrofit2.Retrofit
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
@@ -11,7 +12,7 @@ object NetworkModule {
     val logging = HttpLoggingInterceptor().apply { level = HttpLoggingInterceptor.Level.BODY }
     val client = OkHttpClient.Builder().addInterceptor(logging).build()
 
-    fun provideKakaoApi(context: Context): KakaoApi {
+    fun providSocialLoginApi(context: Context): AuthService {
         val baseUrl = BuildConfig.BASE_URL
         val retrofit = Retrofit.Builder()
             .baseUrl(baseUrl)
@@ -19,7 +20,7 @@ object NetworkModule {
             .addConverterFactory(GsonConverterFactory.create()) // ← Moshi 대신 Gson
             .build()
 
-        return retrofit.create(KakaoApi::class.java)
+        return retrofit.create(AuthService::class.java)
     }
 
 }
